+++ backend/canisters/storage_index/impl/src/lifecycle/post_upgrade.rs
@@ -27,5 +27,6 @@ fn post_upgrade(args: Args) {
 
     init_state(env, data, args.wasm_version);
 
-    info!(version = %args.wasm_version, "Post-upgrade complete");
+    let total_instructions = ic_cdk::api::call_context_instruction_counter();
+    info!(version = %args.wasm_version, total_instructions, "Post-upgrade complete");
 }