+++ backend/canisters/airdrop_bot/c2c_client/src/lib.rs
@@ -1,0 +1,5 @@
+use airdrop_bot_canister::*;
+use canister_client::generate_c2c_call;
+
+// Updates
+generate_c2c_call!(c2c_online_users);