+++ backend/canisters/airdrop_bot/impl/src/guards.rs
@@ -7,3 +7,11 @@ pub fn caller_is_admin() -> Result<(), String> {
         Err("Caller is not an admin".to_string())
     }
 }
+
+pub fn caller_is_online_users_canister() -> Result<(), String> {
+    if read_state(|state| state.is_caller_online_users_canister()) {
+        Ok(())
+    } else {
+        Err("Caller is not the OnlineUsers canister".to_string())
+    }
+}