+++ backend/canisters/online_users/impl/src/lib.rs
@@ -1,3 +1,4 @@
+use crate::model::airdrop_bot_event_batch::AirdropBotEventBatch;
 use crate::model::last_online_dates::LastOnlineDates;
 use crate::model::user_online_minutes::UserOnlineMinutes;
 use canister_state_macros::canister_state;
@@ -8,6 +9,7 @@ use serde::{Deserialize, Serialize};
 use std::cell::RefCell;
 use std::collections::BTreeMap;
 use std::time::Duration;
+use timer_job_queues::GroupedTimerJobQueue;
 use types::{BuildVersion, CanisterId, Cycles, TimestampMillis, Timestamped};
 use utils::env::Environment;
 
@@ -53,10 +55,12 @@ impl RuntimeState {
             git_commit_id: utils::git::git_commit_id().to_string(),
             mark_as_online_count: self.data.mark_as_online_count,
             active_users: self.data.cached_active_users.clone(),
+            sync_online_minutes_to_airdrop_bot_increment: self.data.sync_online_minutes_to_airdrop_bot_increment,
             event_store_client_info,
             stable_memory_sizes: memory::memory_sizes(),
             canister_ids: CanisterIds {
                 user_index: self.data.user_index_canister_id,
+                airdrop_bot: self.data.airdrop_bot_canister_id,
                 event_relay: event_store_canister_id,
                 cycles_dispenser: self.data.cycles_dispenser_canister_id,
             },
@@ -71,19 +75,35 @@ struct Data {
     pub user_online_minutes: UserOnlineMinutes,
     pub principal_to_user_id_map: PrincipalToUserIdMap,
     pub user_index_canister_id: CanisterId,
+    #[serde(default = "CanisterId::anonymous")]
+    pub airdrop_bot_canister_id: CanisterId,
     pub cycles_dispenser_canister_id: CanisterId,
     pub event_store_client: EventStoreClient<CdkRuntime>,
     pub mark_as_online_count: u64,
     pub cached_active_users: ActiveUsers,
+    #[serde(default = "airdrop_bot_event_sync_queue")]
+    pub airdrop_bot_event_sync_queue: GroupedTimerJobQueue<AirdropBotEventBatch>,
+    #[serde(default = "sixty")]
+    pub sync_online_minutes_to_airdrop_bot_increment: u16,
     pub rng_seed: [u8; 32],
     pub test_mode: bool,
+}
+
+fn sixty() -> u16 {
+    60
+}
+
+fn airdrop_bot_event_sync_queue() -> GroupedTimerJobQueue<AirdropBotEventBatch> {
+    GroupedTimerJobQueue::new(1, false)
 }
 
 impl Data {
     pub fn new(
         user_index_canister_id: CanisterId,
+        airdrop_bot_canister_id: CanisterId,
         event_relay_canister_id: CanisterId,
         cycles_dispenser_canister_id: CanisterId,
+        sync_online_minutes_to_airdrop_bot_increment: u16,
         test_mode: bool,
     ) -> Data {
         Data {
@@ -91,12 +111,15 @@ impl Data {
             user_online_minutes: UserOnlineMinutes::default(),
             principal_to_user_id_map: PrincipalToUserIdMap::default(),
             user_index_canister_id,
+            airdrop_bot_canister_id,
             cycles_dispenser_canister_id,
             event_store_client: EventStoreClientBuilder::new(event_relay_canister_id, CdkRuntime::default())
                 .with_flush_delay(Duration::from_secs(60))
                 .build(),
             mark_as_online_count: 0,
             cached_active_users: ActiveUsers::default(),
+            airdrop_bot_event_sync_queue: GroupedTimerJobQueue::new(1, false),
+            sync_online_minutes_to_airdrop_bot_increment,
             rng_seed: [0; 32],
             test_mode,
         }
@@ -113,6 +136,7 @@ pub struct Metrics {
     pub git_commit_id: String,
     pub mark_as_online_count: u64,
     pub active_users: ActiveUsers,
+    pub sync_online_minutes_to_airdrop_bot_increment: u16,
     pub event_store_client_info: EventStoreClientInfo,
     pub stable_memory_sizes: BTreeMap<u8, u64>,
     pub canister_ids: CanisterIds,
@@ -131,6 +155,7 @@ pub struct ActiveUsers {
 #[derive(Serialize, Debug)]
 pub struct CanisterIds {
     pub user_index: CanisterId,
+    pub airdrop_bot: CanisterId,
     pub event_relay: CanisterId,
     pub cycles_dispenser: CanisterId,
 }