+++ backend/tools/canister_installer/src/lib.rs
@@ -170,8 +170,10 @@ async fn install_service_canisters_impl(
     let online_users_canister_wasm = get_canister_wasm(CanisterName::OnlineUsers, version);
     let online_users_init_args = online_users_canister::init::Args {
         user_index_canister_id: canister_ids.user_index,
+        airdrop_bot_canister_id: canister_ids.airdrop_bot,
         event_relay_canister_id: canister_ids.event_relay,
         cycles_dispenser_canister_id: canister_ids.cycles_dispenser,
+        sync_online_minutes_to_airdrop_bot_increment: 60,
         wasm_version: version,
         test_mode,
     };
@@ -194,6 +196,7 @@ async fn install_service_canisters_impl(
         admins: vec![principal],
         user_index_canister_id: canister_ids.user_index,
         local_user_index_canister_id: canister_ids.local_user_index,
+        online_users_canister_id: canister_ids.online_users,
         chat_ledger_canister_id: CHAT_LEDGER_CANISTER_ID,
         wasm_version: version,
         test_mode,