+++ backend/canisters/local_group_index/api/src/updates/c2c_notify_group_index_events.rs
@@ -1,9 +1,0 @@
-use crate::GroupIndexEvent;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Args {
-    pub events: Vec<GroupIndexEvent>,
-}
-
-pub type Response = crate::c2c_group_index::Response;