+++ backend/canisters/local_group_index/impl/src/lifecycle/post_upgrade.rs
@@ -7,7 +7,6 @@ use ic_cdk::post_upgrade;
 use local_group_index_canister::post_upgrade::Args;
 use stable_memory::get_reader;
 use tracing::info;
-use types::BuildVersion;
 use utils::cycles::init_cycles_dispenser_client;
 
 #[post_upgrade]
@@ -16,14 +15,8 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
-
-    for (community_id, community) in data.local_communities.iter() {
-        if community.wasm_version == BuildVersion::new(2, 0, 1618) {
-            data.communities_requiring_upgrade.enqueue((*community_id).into(), false);
-        }
-    }
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);