+++ backend/canisters/notifications/impl/src/updates/c2c_notifications_index.rs
@@ -8,24 +8,11 @@ use stable_memory_map::StableMemoryMap;
 
 #[update(guard = "caller_is_notifications_index", msgpack = true)]
 #[trace]
-fn c2c_sync_index(args: notifications_canister::c2c_sync_index::Args) -> Response {
-    mutate_state(|state| {
-        c2c_sync_index_impl(
-            Args {
-                events: args.events.into_iter().map(|e| e.into()).collect(),
-            },
-            state,
-        )
-    })
-}
-
-#[update(guard = "caller_is_notifications_index", msgpack = true)]
-#[trace]
 fn c2c_notifications_index(args: Args) -> Response {
-    mutate_state(|state| c2c_sync_index_impl(args, state))
+    mutate_state(|state| c2c_notifications_index_impl(args, state))
 }
 
-fn c2c_sync_index_impl(args: Args, state: &mut RuntimeState) -> Response {
+fn c2c_notifications_index_impl(args: Args, state: &mut RuntimeState) -> Response {
     for event in args.events {
         if state.data.idempotency_checker.check(
             state.data.notifications_index_canister_id,