+++ backend/canisters/user/api/src/updates/c2c_notify_group_canister_events.rs
@@ -1,9 +1,0 @@
-use crate::GroupCanisterEvent;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Args {
-    pub events: Vec<GroupCanisterEvent>,
-}
-
-pub type Response = crate::c2c_group_canister::Response;