+++ backend/canisters/user_index/api/src/updates/mod.rs
@@ -6,7 +6,6 @@ pub mod award_external_achievement;
 pub mod c2c_local_user_index;
 pub mod c2c_mark_user_canister_empty;
 pub mod c2c_notify_chit;
-pub mod c2c_notify_events;
 pub mod c2c_notify_low_balance;
 pub mod c2c_report_message;
 pub mod c2c_send_openchat_bot_messages;