+++ backend/canisters/airdrop_bot/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Log total instructions consumed at end of upgrade ([#7551](https://github.com/open-chat-labs/open-chat/pull/7551))
+
 ## [[2.0.1616](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1616-airdrop_bot)] - 2025-02-28
 
 ### Fixed