+++ backend/canisters/airdrop_bot/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Add `min_minutes_online` to airdrop config ([#7563](https://github.com/open-chat-labs/open-chat/pull/7563))
+
 ### Changed
 
 - Log total instructions consumed at end of upgrade ([#7551](https://github.com/open-chat-labs/open-chat/pull/7551))